Add role Provider in cars

+++ Backend/controllers/cars.js
@@ -16,7 +16,7 @@ exports.getCars = async (req, res, next) => {
 
         //Loop over remove fields and delete them from reqQuery
         removeFields.forEach(param=>delete reqQuery[param]);
-        console.log(reqQuery);
+        // console.log(reqQuery);
         
         //Create query string
         let queryStr = JSON.stringify(req.query);
@@ -126,7 +126,15 @@ exports.updateCar = async (req, res, next) => {
             return res.status(400).json({success: false});
         }
 
+        if(req.user.role === 'user' || (car.provider.toString() !== req.user.id && req.user.role === 'provider')) {
+            return res.status(401).json({
+              success:false,
+              message:`User ${req.user.id} is not authorized to update this car`
+            })
+        }
+
         res.status(200).json({success: true, data: car});
+
     } catch(err) {
         console.log(err);
         res.status(400).json({success: false});
@@ -142,6 +150,13 @@ exports.deleteCar = async (req, res, next) => {
 
         if (!car) {
             return res.status(400).json({success: false});
+        }
+
+        if(req.user.role === 'user' || (car.provider.toString() !== req.user.id && req.user.role === 'provider')) {
+            return res.status(401).json({
+              success:false,
+              message:`User ${req.user.id} is not authorized to delete this car`
+            });
         }
 
         await car.deleteOne();
